move types defined in getParamsFromAst to node-sql-parser augment
