+++ typings/node-sql-parser/index.d.ts
@@ -13,4 +13,20 @@ declare module 'node-sql-parser' {
   interface Insert_Replace {
     returning: Returning;
   }
+
+  // types are not provided https://github.com/taozhi8833998/node-sql-parser/issues/1662
+  type Param = { type: 'param'; value: string };
+  type Expr =
+    | {
+        type: 'binary_expr';
+        operator: 'AND' | 'OR';
+        left: Expr;
+        right: Expr;
+      }
+    | {
+        type: 'binary_expr';
+        operator: string;
+        left: ColumnRef | Param;
+        right: ColumnRef | Param;
+      };
 }
